% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicloTurismoApp.R
\name{app_server}
\alias{app_server}
\title{Ciclo de Mobilidade Server}
\usage{
app_server()
}
\value{
function with the call for the Server.
}
\description{
Server (Backend) for the Proof of Concept Ciclo de Mobilidade from 
the city of Recife.

#' @details 
This function only provides the code for the Server (backend)
of the application.
}
