% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicloTurismoApp.R
\name{app_ui}
\alias{app_ui}
\title{Ciclo de Mobilidade UI}
\usage{
app_ui()
}
\value{
the HTML for the frontend.
}
\description{
User Interface (UI) for the Proof of Concept Ciclo de Mobilidade from 
the city of Recife.
}
\details{
This function only provides HTML for the User Interface (frontend)
of the application.
}
