% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntersectionMalhas.R
\name{calcIntersectionMalhas}
\alias{calcIntersectionMalhas}
\title{Calculate intersection between two cycling networks}
\usage{
calcIntersectionMalhas(malhaPermanente, malhaOperacional)
}
\arguments{
\item{malhaPermanente}{data.frame/data.table object with ride information from Strava. 
This is the result from the processing_strava() function.}

\item{malhaOperacional}{data.frame with the geographical features from the 
cycling network. Basically, it is the object resulted from reading the .shp file
using the sf package.}
}
\value{
a data.table object with information from malhaPermanente with the
calculated intersection with malhaOperacional.
}
\description{
This function will calculate the intersection between two cycling networks.
}
\details{
The calculation is done by using a buffering of one of the cycling network (5 meters),
and then calculate the intersection with he other cycling network. #'
}
