% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_strava.R
\name{processing_strava}
\alias{processing_strava}
\title{Processing data from Strava dataset}
\usage{
processing_strava(metadata, shape, dateVar)
}
\value{
a data.frame
}
\description{
This function will process the daily rides dataset files from Strava Metro.
}
\details{
The information provided by Strava Metro is a set of directory containing data 
about all rides in hourly, daily and monthly period.
 This function is aimed to
be used with daily data, that is, files contained in the 'all_edges_daily_*'
 directory.
}
