% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadMalhaOperacional_module.R
\name{uploadMalhaOperacionalInput}
\alias{uploadMalhaOperacionalInput}
\title{Upload Malha Operacional Input}
\usage{
uploadMalhaOperacionalInput(id)
}
\description{
Upload Malha Operacional Input
}
