% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadMalhaPermanente_module.R
\name{uploadMalhaPermanenteInput}
\alias{uploadMalhaPermanenteInput}
\title{Upload Malha Permanente Input}
\usage{
uploadMalhaPermanenteInput(id)
}
\description{
Upload Malha Permanente Input
}
