% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadMalhaPermanente_module.R
\name{uploadMalhaPermanenteServer}
\alias{uploadMalhaPermanenteServer}
\title{Upload Malha Permanente Server}
\usage{
uploadMalhaPermanenteServer(id)
}
\description{
Upload Malha Permanente Server
}
